==> 05_enrich.r <==
dev.off()
rm(list=ls())
dir()
Sys.setenv(LANGUAGE = "en") 
options(stringsAsFactors = FALSE) 
options(clusterProfiler.download.method = "wininet")
library(DOSE)
library(GO.db)
library(org.Hs.eg.db)
library(topGO)
library(GSEABase)
library(clusterProfiler)
library(enrichplot)
library(ggplot2)
library(ggThemeAssist)
library(stringr)
gene1<-read.table("c1.txt",header=F)
gene2<-read.table("c2.txt",header=F)
gene3<-read.table("c3.txt",header=F)
gene4<-read.table("c4.txt",header=F)
gene5<-read.table("c5.txt",header=F)
gene6<-read.table("c6.txt",header=F)
symbol1=as.character(gene1[,1])
symbol2=as.character(gene2[,1])
symbol3=as.character(gene3[,1])
symbol4=as.character(gene4[,1])
symbol5=as.character(gene5[,1])
symbol6=as.character(gene6[,1])
eg1 = bitr(symbol1, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Hs.eg.db")
eg2 = bitr(symbol2, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Hs.eg.db")
eg3 = bitr(symbol3, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Hs.eg.db")
eg4 = bitr(symbol4, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Hs.eg.db")
eg5 = bitr(symbol5, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Hs.eg.db")
eg6 = bitr(symbol6, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Hs.eg.db")
cp = list(C1=eg1$ENTREZID, C2=eg2$ENTREZID,C3=eg3$ENTREZID,C4=eg4$ENTREZID, C5=eg5$ENTREZID,C6=eg6$ENTREZID) 
xx <- compareCluster(cp, fun="enrichKEGG", organism="hsa", pvalueCutoff=0.05, pAdjustMethod = "none", qvalueCutoff = 1)
bb<-dotplot(xx,showCategory=7,title="KEGG pathway", color = "pvalue")+theme_bw()+
  theme(panel.grid.major=element_blank(),panel.grid.minor=element_blank())+
  scale_color_continuous(low="darkorchid3",high = "sienna2")
bb+scale_y_discrete(labels=function(y) str_wrap(y,width=80))
                    
#=============================================================================================================================

dev.off()
rm(list=ls())
dir()
Sys.setenv(LANGUAGE = "en")
options(stringsAsFactors = FALSE)
options(clusterProfiler.download.method = "wininet")
library(DOSE)
library(GO.db)
library(org.Hs.eg.db)
library(topGO)
library(GSEABase)
library(clusterProfiler)
library(enrichplot)
library(ggplot2)
library(ggThemeAssist)
library(stringr)
gene1<-read.table("../data/list.txt",header=F)
symbol1=as.character(gene1[,1])
eg1 = bitr(symbol1, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Hs.eg.db")
kk <- enrichKEGG(gene = eg1$ENTREZID,organism = 'hsa',pvalueCutoff = 0.05,pAdjustMethod = "none", 
                 qvalueCutoff = 1)
c = barplot(kk, showCategory=20) +theme_bw()+theme(panel.grid.major=element_blank(),panel.grid.minor=element_blank())
c+scale_y_discrete(labels=function(y) str_wrap(y,width=80))
